task/username redirection
