+++ pages/[username]/index.tsx
@@ -1,21 +1,17 @@
 import Layout from '../../components/Layout'
 import {
-    DIDxWallet,
     Treasury,
     VerifiableCredentials,
     Defi,
     Headline,
 } from '../../components'
-import { useEffect } from 'react'
-import { useRouter } from 'next/router'
 import { $loading } from '../../src/store/loading'
 import { useStore } from 'effector-react'
-import { $user, updateUser } from '../../src/store/user'
+import { $user } from '../../src/store/user'
 
 function Header() {
     const loading = useStore($loading)
     const user = useStore($user)
-    const Router = useRouter()
     const path = window.location.pathname.toLowerCase()
     const first = path.split('/')[1]
     const username = first.split('.')[0]
@@ -26,20 +22,6 @@ function Header() {
             router: '',
         },
     ]
-
-    useEffect(() => {
-        let domain = ''
-        if (first.split('.')[1] !== undefined) {
-            domain = first.split('.')[1]
-        }
-        updateUser({
-            name: username,
-            domain: domain,
-        })
-        // if (username === 'donate') {
-        //     Router.push('/')
-        // }
-    }, [first, username, Router])
 
     return (
         <>
@@ -60,14 +42,7 @@ function Header() {
                                 ) : username === 'getstarted' ? (
                                     <div />
                                 ) : (
-                                    // ) : username === 'xpoints' ||
-                                    //     user?.name === 'xpoints' ? (
-                                    //     <div />
-                                    // ) : username === 'donate' ? (
-                                    //     <div />
-                                    <DIDxWallet>
-                                        <div />
-                                    </DIDxWallet>
+                                    <></>
                                 )}
                             </>
                         ) : (
